Baje modificaciones del faces-config

+++ src/java/requisiciones/to/TORequisicionDetalle.java
@@ -1,12 +1,13 @@
 package requisiciones.to;
 
 public class TORequisicionDetalle {
+
     private int idRequisicion;
     private int idProducto;
-    private int cantidad;
-    private int cantidadAutorizada;
-    
-     public int getIdRequisicion() {
+    private double cantidad;
+    private double cantidadAutorizada;
+
+    public int getIdRequisicion() {
         return idRequisicion;
     }
 
@@ -22,19 +23,34 @@ public class TORequisicionDetalle {
         this.idProducto = idProducto;
     }
 
-    public int getCantidad() {
+    //    public int getCantidad() {
+    //        return cantidad;
+    //    }
+    //
+    //    public void setCantidad(int cantidad) {
+    //        this.cantidad = cantidad;
+    //    }
+    //
+    //    public int getCantidadAutorizada() {
+    //        return cantidadAutorizada;
+    //    }
+    //
+    //    public void setCantidadAutorizada(int cantidadAutorizada) {
+    //        this.cantidadAutorizada = cantidadAutorizada;
+    //    }
+    public double getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
 
-    public int getCantidadAutorizada() {
+    public double getCantidadAutorizada() {
         return cantidadAutorizada;
     }
 
-    public void setCantidadAutorizada(int cantidadAutorizada) {
+    public void setCantidadAutorizada(double cantidadAutorizada) {
         this.cantidadAutorizada = cantidadAutorizada;
     }
 }
